+++ SDL_project_0/ComponentTransform.h
@@ -12,12 +12,12 @@ protected:
 public:
 
 	ComponentTransform& SetTransform(Transform& t);
-	ComponentTransform& SetPosition(Vector2i& pos);
+	ComponentTransform& SetPosition(Vector2d& pos);
 	ComponentTransform& SetRotation(double rot);
 	ComponentTransform& SetScale(Vector2f& sc);
 
 	ComponentTransform& SetTransform(Transform&& t) { return SetTransform(t); }
-	ComponentTransform& SetPosition(Vector2i&& pos) { return SetPosition(pos); }
+	ComponentTransform& SetPosition(Vector2d&& pos) { return SetPosition(pos); }
 	ComponentTransform& SetScale(Vector2f&& sc) { return SetScale(sc); }
 
 
@@ -25,7 +25,7 @@ public:
 
 	virtual Transform GetWorldTransform();
 	virtual Transform& GetLocalTransform();
-	virtual Vector2i GetPosition();
+	virtual Vector2d GetPosition();
 	virtual double GetRotation();
 	virtual Vector2f GetScale();
 
@@ -33,4 +33,3 @@ public:
 
 	void OnSpawn() override;
 };
-
